refactor: improve code style and readability
Replace modulo operator with is_multiple_of method for clearer
intent across multiple files. Reformat long lines to respect72
character limit and improve readability of complex expressions.
Sort public use statements alphabetically. Add numeric separators
to long floating point constants for better readability. Simplify
conditional logic using && operator in progress checks. Fix array
indexing calculation in spectral.rs and remove unnecessary type
casts in multigrid.rs.

+++ src/tooling/core/poisson/multigrid.rs
@@ -491,7 +491,7 @@ impl PoissonSolver for Multigrid {
         let max_iter = 100u64;
         for _iter in 0..max_iter {
             if let Some(ref p) = self.progress {
-                p.set_intra_progress(_iter as u64, max_iter);
+                p.set_intra_progress(_iter, max_iter);
             }
             v_cycle(
                 &mut phi,
